requirements.txt
dspy-ai==2.5.20
openai==1.53.0
pandas==2.2.3
ujson==5.10.0
ragas==0.2.3
nltk==3.9.1
